requirements.txt
python-dotenv>=1.0.0
phonemizer>=3.2.1
torch>=2.0.1
transformers>=4.30.2
scipy>=1.10.1
munch>=2.5.0
soundfile>=0.12.1
nltk>=3.8.1
google-genai>=1.3.0